fix: provide graalvm for both old and new version of aws sdk otel
